Fix endpoints
