+++ hub/backend/app/api/main.py
@@ -8,6 +8,7 @@ from app.api.routes import (
     misc,
     orgs,
     projects,
+    references,
     users,
 )
 from app.api.routes.projects import releases
@@ -18,6 +19,7 @@ api_router.include_router(login.router, tags=["login"])
 api_router.include_router(users.router, tags=["users"])
 api_router.include_router(misc.router, tags=["misc"])
 api_router.include_router(projects.router, tags=["projects"])
+api_router.include_router(references.router, tags=["references"])
 api_router.include_router(releases.router, tags=["releases"])
 api_router.include_router(orgs.router, tags=["orgs"])
 api_router.include_router(datasets.router, tags=["datasets"])
