+++ hub/frontend/src/client/index.ts
@@ -10,6 +10,7 @@ export {
   type Options,
   OrgsService,
   ProjectsService,
+  ReferencesService,
   ReleasesService,
   UsersService,
 } from "./sdk.gen"
@@ -437,6 +438,11 @@ export type {
   GetProjectZoteroLibrariesErrors,
   GetProjectZoteroLibrariesResponse,
   GetProjectZoteroLibrariesResponses,
+  GetReferencesData,
+  GetReferencesError,
+  GetReferencesErrors,
+  GetReferencesResponse,
+  GetReferencesResponses,
   GetReleaseCommentsData,
   GetReleaseCommentsError,
   GetReleaseCommentsErrors,
@@ -490,11 +496,6 @@ export type {
   GetUserOverleafTokenData,
   GetUserOverleafTokenResponse,
   GetUserOverleafTokenResponses,
-  GetUserReferencesData,
-  GetUserReferencesError,
-  GetUserReferencesErrors,
-  GetUserReferencesResponse,
-  GetUserReferencesResponses,
   GetUserStorageData,
   GetUserStorageResponse,
   GetUserStorageResponses,
